ton-vs-daily-stat-bot/requirements.txt
python-telegram-bot==13.10
psycopg2-binary==2.9.3
matplotlib==3.5.1
